+++ src/aks-preview/azext_aks_preview/managed_cluster_decorator.py
@@ -2852,6 +2852,22 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc.http_proxy_config = self.context.get_http_proxy_config()
         return mc
 
+    def update_kube_proxy_config(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update kube proxy config for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        if not mc.network_profile:
+            raise UnknownError(
+                "Unexpectedly get an empty network profile in the process of updating kube-proxy config."
+            )
+
+        mc.network_profile.kube_proxy_config = self.context.get_kube_proxy_config()
+
+        return mc
+
     def update_pod_security_policy(self, mc: ManagedCluster) -> ManagedCluster:
         """Update pod security policy for the ManagedCluster object.
 
@@ -3228,5 +3244,7 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_linux_profile(mc)
         # update outbound type
         mc = self.update_outbound_type_in_network_profile(mc)
+        # update kube proxy config
+        mc = self.update_kube_proxy_config(mc)
 
         return mc