+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,26 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+0.5.118
++++++++
+
+* Support enabling syslog collection in monitoring on AKS clusters with msi auth
+* Add `--enable-syslog` option in aks create and aks enable-addons
+
+0.5.117
++++++++
+
+* Add custom transform for custom CA
+* Support updating kube-proxy configuration with `az aks update --kube-proxy-config file.json`.
+
+0.5.116
++++++++
+
 * Fix `az aks update` command failing on updating the ssh key value if cluster was created without ssh key, see issue `\#5559 <https://github.com/Azure/azure-cli-extensions/issues/5559>`_.
-* Mark "--enable-pod-security-policy" deprecated
-* Deny create request if binding existed for command "trustedaccess rolebinding create"
-* Support AAD clusters for "az aks kollect"
+* Mark "--enable-pod-security-policy" deprecated.
+* Deny create request if binding existed for command "trustedaccess rolebinding create".
+* Support AAD clusters for "az aks kollect".
+* Vendor new SDK and bump API version to 2022-10-02-preview.
 
 0.5.115
 +++++++